# 1-todo-entity-svc/go/model_todo_response.go
/*
 * Todo API
 *
 * A simple Todo API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type TodoResponse struct {
	Id int32 `json:"id"`

	UserId string `json:"userId"`

	Title string `json:"title"`

	Description string `json:"description"`

	Completed bool `json:"completed"`
}